Move beneficiary related actions to be under beneficiaries tab
