+++ src/components/BenefitPlanTabPanel.js
@@ -1,23 +1,17 @@
 import React, { useState } from 'react';
 import {
-  Paper, Grid, Button, Tooltip,
+  Paper, Grid,
 } from '@material-ui/core';
 import { injectIntl } from 'react-intl';
 import {
-  Contributions, useHistory, useModulesManager, useTranslations,
+  Contributions,
 } from '@openimis/fe-core';
 import { makeStyles } from '@material-ui/core/styles';
 import {
   BENEFIT_PLAN_BENEFICIARIES_TAB_WRAPPER_VALUE,
   BENEFIT_PLAN_TABS_LABEL_CONTRIBUTION_KEY,
   BENEFIT_PLAN_TABS_PANEL_CONTRIBUTION_KEY,
-  DEDUPLICATION_SELECT_FIELD_DIALOG_CONTRIBUTION_KEY,
-  MODULE_NAME,
-  PAYROLL_CREATE_RIGHTS_PUB_REF,
-  PAYROLL_PAYROLL_ROUTE,
 } from '../constants';
-import BenefitPlanBeneficiariesUploadDialog from '../dialogs/BenefitPlanBeneficiariesUploadDialog';
-import BenefitPlanBeneficiariesUploadHistoryDialog from '../dialogs/BenefitPlanBeneficiariesUploadHistoryDialog';
 
 const useStyles = makeStyles((theme) => ({
   paper: theme.paper.paper,
@@ -44,9 +38,6 @@ function BenefitPlanTabPanel({
   intl, rights, benefitPlan, setConfirmedAction, onActiveTabChange,
 }) {
   const classes = useStyles();
-  const modulesManager = useModulesManager();
-  const { formatMessage } = useTranslations(MODULE_NAME, modulesManager);
-  const history = useHistory();
   const [activeTab, setActiveTab] = useState(BENEFIT_PLAN_BENEFICIARIES_TAB_WRAPPER_VALUE);
 
   const isSelected = (tab) => tab === activeTab;
@@ -58,66 +49,18 @@ function BenefitPlanTabPanel({
     onActiveTabChange(tab);
   };
 
-  const payrollCreateRights = modulesManager.getRef(PAYROLL_CREATE_RIGHTS_PUB_REF);
-
-  const handleCreatePayrollButton = () => {
-    history.push(
-      `/${modulesManager.getRef(PAYROLL_PAYROLL_ROUTE)}/null/null/${benefitPlan.id}`,
-    );
-  };
-
   return (
     <Paper className={classes.paper}>
       <Grid container className={`${classes.tableTitle} ${classes.tabs}`}>
-        <div style={{ width: '100%' }}>
-          <div style={{ float: 'left' }}>
-            <Contributions
-              contributionKey={BENEFIT_PLAN_TABS_LABEL_CONTRIBUTION_KEY}
-              intl={intl}
-              rights={rights}
-              value={activeTab}
-              onChange={handleChange}
-              isSelected={isSelected}
-              tabStyle={tabStyle}
-            />
-          </div>
-          <div style={{ float: 'right', paddingRight: '16px' }}>
-            {rights.includes(payrollCreateRights) && (
-            <Tooltip
-              title={formatMessage('benefitPlan.benefitPlanTabPanel.createPayroll.tooltip')}
-              disableHoverListener={benefitPlan?.hasPaymentPlans}
-            >
-              <span>
-                <Button
-                  onClick={handleCreatePayrollButton}
-                  variant="outlined"
-                  color="#DFEDEF"
-                  disabled={!benefitPlan?.hasPaymentPlans}
-                  className={classes.button}
-                  style={{
-                    border: '0px',
-                    marginTop: '6px',
-                  }}
-                >
-                  {formatMessage('benefitPlan.benefitPlanTabPanel.createPayroll')}
-                </Button>
-              </span>
-            </Tooltip>
-            )}
-
-            <Contributions
-              contributionKey={DEDUPLICATION_SELECT_FIELD_DIALOG_CONTRIBUTION_KEY}
-              intl={intl}
-              benefitPlan={benefitPlan}
-            />
-            <BenefitPlanBeneficiariesUploadDialog
-              benefitPlan={benefitPlan}
-            />
-            <BenefitPlanBeneficiariesUploadHistoryDialog
-              benefitPlan={benefitPlan}
-            />
-          </div>
-        </div>
+        <Contributions
+          contributionKey={BENEFIT_PLAN_TABS_LABEL_CONTRIBUTION_KEY}
+          intl={intl}
+          rights={rights}
+          value={activeTab}
+          onChange={handleChange}
+          isSelected={isSelected}
+          tabStyle={tabStyle}
+        />
       </Grid>
       <Contributions
         contributionKey={BENEFIT_PLAN_TABS_PANEL_CONTRIBUTION_KEY}
